requirements.txt
torch==2.1.2
torchvision==0.16.2
segmentation-models-pytorch==0.3.3
timm==0.9.12
einops==0.7.0
albumentations==1.3.1
opencv-python-headless==4.9.0.80
scikit-learn==1.4.0
matplotlib==3.8.2
numpy==1.26.3
Pillow==10.2.0
PyYAML==6.0.1
tqdm==4.66.1
tensorboard==2.15.1
gradio==4.14.0
gdown==5.1.0
